refactor(p1am-hmi): split TemperatureControl; fold recall onto seedDraftText
TemperatureControl.tsx was 1975 lines — 475 over DEFAULT_MAX_SOURCE_LINES —
so fleet-fast-guardrails rejected any commit touching it. The heater screen
was the one operator surface that could not be corrected without first being
restructured.

Following the TuningPanel shape (#4053), the container keeps all state, the
rolling trend buffer and every /api/temperature/* call; the sections are
prop-driven components:

  TemperatureControl.tsx      1975 -> 717
  TemperatureTrend.tsx             680  (SVG trend + its own view state)
  ThermocoupleSelector.tsx         196
  TemperatureConfigPanel.tsx       148
  TemperatureStatusHeader.tsx      136
  HeaterStartStopButton.tsx         55
  lib/temperatureTrend.ts          171  (pure sample/readout math)

The Start/Stop command button existed as two byte-identical copies on the same
screen; on a control that energizes a heater those could come to disagree about
whether a command is safe to send. One component now serves both, differing
only by an appended CSS class.

DRY cleanup #4053 had to defer: recallSetpointText now returns
seedDraftText(lastSetpointC, operatorTouched, 1) rather than carrying its own
copy of the operator-ownership rule. The heater domain types moved to
src/types.ts, removing the cycle where useTelemetryStream imported
TemperatureStatus from a component.

No behaviour change: every moved block is verbatim apart from prop renames and
JSX whitespace that collapses identically. lint (0 errors, same 2 pre-existing
warnings), build and 444/444 vitest tests pass.
